requirements.txt
langchain-core
langchain-openai
langchain_community
llama-index
pydantic
s3fs